[S12P21E206-89] Feat: AWS S3 + 판매자 펀딩 수정, 리팩토링 (PUT -> PATCH)

+++ back/funding/src/main/java/com/ssafy/funding/service/impl/S3FileService.java
@@ -50,11 +50,8 @@ public class S3FileService {
         if (fileUrl == null || fileUrl.isBlank()) return;
 
         String key = extractKeyFromUrl(fileUrl);
-        if (amazonS3.doesObjectExist(bucket, key)) {
-            amazonS3.deleteObject(bucket, key);
-        } else {
-            throw new CustomException(FAIL_FILE_DELETE);
-        }
+
+        if (amazonS3.doesObjectExist(bucket, key)) amazonS3.deleteObject(bucket, key);
     }
     public void deleteFiles(List<String> fileUrls) {
         if (fileUrls == null || fileUrls.isEmpty()) return;
